This is game is developed is for entertainment purpose , whole game is developed in core java and JDBC technology
I ,Nitesh Agarwal, made this game during summer training 2014 with my
teammate Chiranjeev Chunhan. we developed this game for  making our
logic parts Strong in programming as well as we always had a curiosity
that how a game is made . DOWNLOAD it,  it is very   interesting, that
You will be enjoying feel of  shooting enemies.